test(readUserSettings): complete testing for readUserSettings
Much of this commit was exploratory for finding ways to interact with the test workspace while running a test suite.  This test changes settings at an arbitrary point in the testing process and so any tests that rely on settings should have those settings mocked rather than reading from the workspace.